fix: prevent deleting AWS OIDC integration used by EAS
The External Audit Storage feature for Teleport Cloud depends on an AWS
OIDC integration to exist in order to get credentials for the customer
AWS account.
It's currently possible for a customer to delete the AWS OIDC
integration referenced by the EAS configuration, leading to the EAS
integration breaking and the Auth service crashing the next time it's
started up.

This commit prevents deleting an AWS OIDC integration currently
referenced by and EAS integration, and prevents creating an EAS
integration referencing an unknown AWS OIDC integration.

+++ lib/services/local/integrations.go
@@ -36,11 +36,34 @@ const (
 
 // IntegrationsService manages Integrations in the Backend.
 type IntegrationsService struct {
-	svc generic.Service[types.Integration]
+	svc              generic.Service[types.Integration]
+	easSvc           *ExternalAuditStorageService
+	backend          backend.Backend
+	deleteAllEnabled bool
+}
+
+// IntegrationsServiceOption is a functional option for the IntegrationsService.
+type IntegrationsServiceOption func(*IntegrationsService)
+
+// WithDeleteAllIntegrationsEnabled configure the IntegrationsService to support the DeleteAllIntegrations
+// method, which does not include protections against deleting integrations referenced by other components and
+// should only be used in e.g. a local cache.
+func WithDeleteAllIntegrationsEnabled(deleteAllEnabled bool) func(*IntegrationsService) {
+	return func(svc *IntegrationsService) {
+		svc.deleteAllEnabled = deleteAllEnabled
+	}
+}
+
+// WithExternalAuditStorageService configure the IntegrationsService to use the given
+// ExternalAuditStorageService.
+func WithExternalAuditStorageService(easSvc *ExternalAuditStorageService) func(*IntegrationsService) {
+	return func(svc *IntegrationsService) {
+		svc.easSvc = easSvc
+	}
 }
 
 // NewIntegrationsService creates a new IntegrationsService.
-func NewIntegrationsService(backend backend.Backend) (*IntegrationsService, error) {
+func NewIntegrationsService(backend backend.Backend, opts ...IntegrationsServiceOption) (*IntegrationsService, error) {
 	svc, err := generic.NewService(&generic.ServiceConfig[types.Integration]{
 		Backend:       backend,
 		PageLimit:     defaults.MaxIterationLimit,
@@ -52,10 +75,20 @@ func NewIntegrationsService(backend backend.Backend) (*IntegrationsService, erro
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
-
-	return &IntegrationsService{
-		svc: *svc,
-	}, nil
+	integrationsSvc := &IntegrationsService{
+		svc:     *svc,
+		backend: backend,
+	}
+	for _, opt := range opts {
+		opt(integrationsSvc)
+	}
+	if integrationsSvc.easSvc == nil {
+		integrationsSvc.easSvc, err = NewExternalAuditStorageServiceFallible(backend, WithIntegrationsService(integrationsSvc))
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+	}
+	return integrationsSvc, nil
 }
 
 // ListIntegrations returns a paginated list of Integration resources.
@@ -100,10 +133,51 @@ func (s *IntegrationsService) UpdateIntegration(ctx context.Context, ig types.In
 
 // DeleteIntegration removes the specified Integration resource.
 func (s *IntegrationsService) DeleteIntegration(ctx context.Context, name string) error {
-	return trace.Wrap(s.svc.DeleteResource(ctx, name))
+	ig, err := s.svc.GetResource(ctx, name)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+
+	awsOIDCSpec := ig.GetAWSOIDCIntegrationSpec()
+	if awsOIDCSpec == nil {
+		// Not an AWS OIDC integration, go ahead and delete it.
+		return trace.Wrap(s.svc.DeleteResource(ctx, name))
+	}
+
+	// Avoid deleting AWS OIDC integrations referenced by any External Audit Storage integration.
+	// It's necessary to do this under the same lock that is used to create and promote EAS configurations to
+	// avoid a race condition where an AWS OIDC configuration could be deleted immediately after a new EAS
+	// configuration is added.
+	err = backend.RunWhileLocked(ctx, backend.RunWhileLockedConfig{
+		LockConfiguration: backend.LockConfiguration{
+			Backend:  s.backend,
+			LockName: externalAuditStorageLockName,
+			TTL:      externalAuditStorageLockTTL,
+		},
+	}, func(ctx context.Context) error {
+		if draft, err := s.easSvc.GetDraftExternalAuditStorage(ctx); err == nil {
+			if draft.Spec.IntegrationName == name {
+				return trace.BadParameter("cannot delete AWS OIDC integration currently referenced by draft External Audit Storage integration")
+			}
+		} else if !trace.IsNotFound(err) {
+			return trace.Wrap(err)
+		}
+		if cluster, err := s.easSvc.GetClusterExternalAuditStorage(ctx); err == nil {
+			if cluster.Spec.IntegrationName == name {
+				return trace.BadParameter("cannot delete AWS OIDC integration currently referenced by cluster External Audit Storage integration")
+			}
+		} else if !trace.IsNotFound(err) {
+			return trace.Wrap(err)
+		}
+		return trace.Wrap(s.svc.DeleteResource(ctx, name))
+	})
+	return trace.Wrap(err)
 }
 
 // DeleteAllIntegrations removes all Integration resources.
 func (s *IntegrationsService) DeleteAllIntegrations(ctx context.Context) error {
+	if !s.deleteAllEnabled {
+		return trace.BadParameter("Deleting all integrations is not supported, this is a bug")
+	}
 	return trace.Wrap(s.svc.DeleteAllResources(ctx))
 }
